tweaks

+++ crates/multiemu/src/component/schedulable.rs
@@ -2,5 +2,5 @@ use num::rational::Ratio;
 use super::Component;
 
 pub trait SchedulableComponent: Component {
-    fn run(&self, period: u32);
+    fn run(&self, period: u64);
 }
